chore: rename old accordion files to prevent naming conflict
